requirements.txt
django
djangorestframework
django-redis
pytest
pytest-django
Faker
gunicorn
psycopg2-binary
django-prometheus
djangorestframework-simplejwt
kafka-python
